tools/jira-monitor/requirements.txt
requests
certifi
pyyaml
python-dotenv
python-slugify
